+++ docs/changelog.rst
@@ -6,6 +6,7 @@ v0.6.4
 
 * Added creation date to series
 * Added library, releases, bookmarks RSS feeds
+* Added support for custom chapter names
 
 v0.6.3
 ^^^^^^